ERMP-9 Implement a possibility to view movie details Given: the user is on the main page And: there are movies in the list When: the user clicks on a movie card in the list Then: the search panel at the top is replaced with clicked movie details
