+++ src/utils/const/moviList.ts
@@ -1,86 +1,3 @@
-import { IDropDownItem } from 'components/DropDown/DropDownItem/DropDownItem';
-
-export const ADD_MOVE_BUTTON = {
-  text: '+ add move',
-  classNames: ['add__button', 'upper__text']
-};
-
-export const CONGRATULATIONS_ADD_MOVIE = {
-  headline: 'congratulations !',
-  text: 'The movie has been added to database successfully '
-};
-
-export const CONFIRM_BUTTON = {
-  text: 'confirm',
-  classNames: ['confirm__button', 'upper__text']
-};
-
-export const FORM_MOVIE_RESET_BUTTON = {
-  text: 'reset',
-  classNames: ['form__movie__button', 'reset__button', 'upper__text']
-};
-
-export const FORM_MOVIE_SUBMIT_BUTTON = {
-  text: 'submit',
-  classNames: ['form__movie__button', 'submit__button', 'upper__text']
-};
-
-export const SEARCH_COMPONENT = {
-  headline: 'find you movie',
-  placeholder: 'What do you want to watch?',
-  button: {
-    text: 'search',
-    classNames: ['search__button', 'upper__text']
-  }
-};
-
-export const CONFIRM_COMPONENT = {
-  DELETE_ACTION: {
-    headline: 'Delete MOVIE',
-    confirmText: 'Are you sure you want to delete this movie?'
-  }
-};
-
-export const GENRE_LIST = [
-  'Fantasy',
-  'Adventure',
-  'Science Fiction',
-  'Drama',
-  'Romance'
-];
-
-export const GENRE_BAR = [
-  { label: 'all' },
-  ...GENRE_LIST.map(genreName => ({ label: genreName }))
-] as Array<IDropDownItem>;
-
-export const SORT_BAR = {
-  labelText: 'sort by',
-  list: [
-    {
-      label: 'release date'
-    },
-    {
-      label: 'name'
-    },
-    {
-      label: 'genre'
-    },
-    {
-      label: 'rating'
-    }
-  ] as Array<IDropDownItem>
-};
-
-export const MOVIE_ACTION = [
-  {
-    label: 'edit'
-  },
-  {
-    label: 'delete'
-  }
-] as Array<IDropDownItem>;
-
 export const MOVIE_LIST = [
   {
     id: 337167,
